+++ omniledger/darc/darc_test.go
@@ -6,7 +6,7 @@ import (
 	"testing"
 	//	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	"gopkg.in/dedis/onet.v1/log"
+	"gopkg.in/dedis/onet.v2/log"
 )
 
 func TestDarc(t *testing.T) {
