+++ omniledger/service/struct.go
@@ -6,7 +6,14 @@ import (
 
 	bolt "github.com/coreos/bbolt"
 	"github.com/dedis/student_18_omniledger/omniledger/collection"
+	"github.com/dedis/student_18_omniledger/omniledger/darc"
+	"gopkg.in/dedis/onet.v2/network"
 )
+
+func init() {
+	network.RegisterMessages(&Transaction{},
+		&darc.Signature{})
+}
 
 type collectionDB struct {
 	db         *bolt.DB
@@ -105,4 +112,23 @@ func (c *collectionDB) GetValue(key []byte) (value, sig []byte, err error) {
 // RootHash returns the hash of the root node in the merkle tree.
 func (c *collectionDB) RootHash() []byte {
 	return c.coll.GetRoot()
+}
+
+// Transaction is the struct specifying the modifications to the skipchain.
+// Key is the key chosen by the user, Kind is the kind of value to store
+// (e.g. a drac...). The key used in the conode's collection will be
+// Kind ':' Key, in order to maintain key uniqueness across different kinds
+// of values.
+// For a Transaction to be valid, there must exist a path from the master-darc
+// in the genesis block to the SubjectPK in Signature.
+type Transaction struct {
+	Key   []byte
+	Kind  []byte
+	Value []byte
+	// type Signature struct {
+	//     Signature []byte
+	//     Signer SubjectPK
+	// }
+	// The signature is performed on the concatenation of the []bytes
+	Signature darc.Signature
 }
